requirements.txt
setuptools
apache-airflow